requirements.txt
uvicorn
fastapi
pydantic
motor
python-jose
passlib
groq
easyocr
PyMuPDF
numpy
pdfplumber
firebase_admin
pillow
python-dotenv
bson